date: allow 10+ char regex
Inputs were limited to 10 max chars, or else the regex would fail.

This change allows something like: `2023-01-25T10:00:00Z` to be
parsed into `2023-01-25` where as before it would fail because
it was longer than 10 chars.

In the regex, this means using `.*$` at the end instead of `$`
right on the 10th character, to allow matching an infinitely
long string.